pkg_managers: gomod: Parse the toolchain keyword from go.mod
Unfortunately, with Go 1.21 it's not enough to rely solely on what the
mod file declares with the 'go' line, but also what it declares
with the 'toolchain' line. While 'toolchain' is only suggestive,
i.e. not usually enforced, the situation changes when combined with
GOTOOLCHAIN=auto. Start parsing that keyword for future use.

Inspired-by: https://github.com/containerbuildsystem/cachi2/commit/2c83a586eab87487b1f2303f0c037748b2603059

Signed-off-by: Erik Skultety <eskultet@redhat.com>