pkg_managers: gomod: Enable Full Go 1.21 support
Commit 9f08f8e8 added preliminary support for Go 1.21 only, meaning
that we'd use a fixed toolchain version of 1.21.0 (or the bundled Go
toolchain, whichever was bigger). Some projects do bump even the micro
version of their required Go toolchain which means that we'd be
forced to refresh the container image with every Go micro release to
keep up with these bumps for the dependency prefetching (looking at you
Openshift).

Starting with Go 1.21, however, one is able to instruct Go to download
a newer toolchain automatically if the installed one doesn't comply
with the requirement in go.mod. That behaviour is controller by the
GOTOOLCHAIN variable which will need to be set as 'GOTOOLCHAIN=auto' to
override the default behaviour of Go (depending on the installation)
which always prefers the installed toolchain.

One thing to consider is the fact that a newer toolchain never
downloads an older one, meaning that if the project's build image
doesn't supply the desired version of Go (either with through the 'go'
line or the 'toolchain' line) and the consumer overrides the
GOTOOLCHAIN setting to 'auto', the build will fail in either case
because the correct toolchain version won't be available.
To mitigate this on our side we'll always base our processing on the
0th release of a minor Go version (i.e. 1.21.0, 1.22.0, etc.) which
ensures that with GOTOOLCHAIN=auto Go will always download and cache
any newer toolchain from any dependent module.

Note that we're only going to set GOTOOLCHAIN=auto for the dependency
fetching, i.e. in our own environment. Consumers are free
to override the GOTOOLCHAIN at their will in their build environment.
With that in mind, we're also dropping the GOTOOLCHAIN=local override
we have been forcing upon the users so far for the very same reason.

Inspired-by: https://github.com/containerbuildsystem/cachi2/commit/e0732c78f7df24de2b2a6af0750e40fa2b6396a1

Signed-off-by: Erik Skultety <eskultet@redhat.com>